backend/requirements.txt
# AP2 Travel Agent Backend Dependencies

# FastAPI framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Data validation
pydantic==2.9.0

# Ollama client for local LLM
ollama==0.3.3

# HTTP client for A2A communication
httpx==0.27.2

# Environment management
python-dotenv==1.0.1

# File handling
python-multipart==0.0.12
aiofiles==24.1.0

# Server utilities
starlette==0.38.6

# Testing (optional)
pytest==8.3.3
pytest-asyncio==0.24.0
